fix: 회원가입 화면 수정 - 주소 입력 시 팝업이 안꺼지는 문제 해결 - users 폴더 내부 구조 변경
