+++ src/main/java/com/pyeonrimium/queuing/reservation/controller/ReservationController.java
@@ -16,7 +16,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
 
 import com.pyeonrimium.queuing.reservation.Service.ReservationService;
 import com.pyeonrimium.queuing.reservation.domains.ReservationDeleteResponse;
