--- src/main/java/com/pyeonrimium/queuing/users/controllers/Find_passwordRequest.java
@@ -1,13 +1,0 @@
-package com.pyeonrimium.queuing.users.controllers;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class Find_passwordRequest {
-	
-	private String id;
-	private String name;
-	private String phone;
-}
