+++ src/main/java/com/pyeonrimium/queuing/users/domains/dtos/Find_passwordRequest.java
@@ -1,15 +1,13 @@
-package com.pyeonrimium.queuing.users.controllers;
+package com.pyeonrimium.queuing.users.domains.dtos;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class SignupRequest {
+public class Find_passwordRequest {
 	
 	private String id;
-	private String password;
 	private String name;
-	private String address;
 	private String phone;
 }
