+++ src/main/webapp/resources/css/signupStyles.css
@@ -1,15 +1,15 @@
-
 body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-  height: 100vh;
-  background-color: #f5f5f5;
+	font-family: Arial, sans-serif;
+	margin: 0;
+	padding: 0;
+	height: 100vh;
+	background-color: #f5f5f5;
 }
 
 .container {
-    flex-direction: column;
-    min-height: 100vh;
+	display: flex;
+	flex-direction: column;
+	min-height: 100vh;
 }
 
 main {
@@ -20,55 +20,71 @@ main {
 }
 
 .signup-container {
-  margin: auto;
-  flex: 1;
-  width: 100%;
-  max-width: 400px;
-  background-color: #fff;
-  padding: 20px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
+	margin: auto;
+	flex: 1;
+	width: 100%;
+	max-width: 600px;
+	background-color: #fff;
+	padding: 30px;
+	box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+	border-radius: 8px;
 }
 
 h2 {
-  margin-top: 0;
-  font-size: 24px;
+	margin-top: 0;
+	font-size: 24px;
 }
 
 .input-group {
-  margin-bottom: 15px;
+	margin-bottom: 15px;
+	display: flex;
 }
 
 label {
-  font-weight: bold;
-  display: block;
+	width: 110px;
+	margin-right: 20px;
+	font-weight: bold;
+	display: flex;
+	align-items: center;
 }
 
-input[type="text"],
-input[type="password"] {
-  width: 100%;
-  padding: 8px;
-  box-sizing: border-box;
+input {
+	flex-grow: 1;
+	padding: 8px;
+	box-sizing: border-box;
 }
 
-button {
-  width: 100%;
-  padding: 10px;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  cursor: pointer;
-  font-size: 16px;
-  margin-top: 10px;
-  border-radius: 4px;
+#addrBtn {
+	padding: 10px;
+	background-color: #007bff;
+	color: white;
+	border: none;
+	cursor: pointer;
+	margin-left: 10px;
+	border-radius: 4px;
+}
+
+.buttons {
+	margin-top: 30px;
+}
+
+#submitBtn {
+	width: 100%;
+	padding: 10px;
+	background-color: #007bff;
+	color: white;
+	border: none;
+	cursor: pointer;
+	font-size: 16px;
+	border-radius: 4px;
 }
 
 button:hover {
-  background-color: #0056b3;
+	background-color: #0056b3;
 }
 
 .error-message {
-  color: red;
-  display: none;
-  margin-top: 10px;
+	color: red;
+	display: none;
+	margin-top: 10px;
 }
